Moving to src/platform/plugins/shared
