+++ src/platform/plugins/shared/metrics_experience/kibana.jsonc
