+++ src/platform/plugins/shared/metrics_experience/tsconfig.json
