+++ tsconfig.base.json
@@ -1322,8 +1322,8 @@
       "@kbn/metrics-config/*": ["src/platform/packages/private/opentelemetry/kbn-metrics-config/*"],
       "@kbn/metrics-data-access-plugin": ["x-pack/solutions/observability/plugins/metrics_data_access"],
       "@kbn/metrics-data-access-plugin/*": ["x-pack/solutions/observability/plugins/metrics_data_access/*"],
-      "@kbn/metrics-experience-plugin": ["x-pack/platform/plugins/shared/metrics_experience"],
-      "@kbn/metrics-experience-plugin/*": ["x-pack/platform/plugins/shared/metrics_experience/*"],
+      "@kbn/metrics-experience-plugin": ["src/platform/plugins/shared/metrics_experience"],
+      "@kbn/metrics-experience-plugin/*": ["src/platform/plugins/shared/metrics_experience/*"],
       "@kbn/ml-agg-utils": ["x-pack/platform/packages/private/ml/agg_utils"],
       "@kbn/ml-agg-utils/*": ["x-pack/platform/packages/private/ml/agg_utils/*"],
       "@kbn/ml-anomaly-utils": ["x-pack/platform/packages/shared/ml/anomaly_utils"],
