--- x-pack/platform/plugins/shared/metrics_experience/jest.config.js
@@ -1,14 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-const path = require('path');
-
-module.exports = {
-  preset: '@kbn/test',
-  rootDir: path.resolve(__dirname, '../../../../..'),
-  roots: ['<rootDir>/x-pack/platform/plugins/shared/apm_sources_access'],
-};
